Proposed change to the Distinguished Name syntax
Using ANY for the value still allows us to use octet strings, but allows
whoever defined the OID to determine what the syntax ought to be. As
this is a Distinguished Name, that will be a human-readable string most
of the time, but we want to explicitly allow NULL for use with the
Authority OID in the target address field.